requirements.txt
pandas
serpapi
python-dotenv
google-search-results
